requirements.txt
python-telegram-bot==20.7
discord.py==2.3.2
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
json5==0.9.14
websockets==12.0
tweepy>=4.12.0
beautifulsoup4==4.12.2
selenium==4.15.2
schedule==1.2.0
requests-oauthlib==1.3.1
